fix(app): Changes acc to review suggestion

+++ apps/evm_family/evm_helpers.c
@@ -125,7 +125,7 @@ static inline bool is_account_hd_path(const uint32_t *path, uint32_t depth);
  * @return bool value indicating hash calculation status.
  */
 static bool evm_get_personal_data_digest(uint8_t *msg_data,
-                                         uint16_t msg_data_size,
+                                         uint32_t msg_data_size,
                                          uint8_t *digest_out);
 /*****************************************************************************
  * STATIC VARIABLES
@@ -158,7 +158,7 @@ static inline bool is_account_hd_path(const uint32_t *path, uint32_t depth) {
 }
 
 static bool evm_get_personal_data_digest(uint8_t *msg_data,
-                                         uint16_t msg_data_size,
+                                         uint32_t msg_data_size,
                                          uint8_t *digest_out) {
   if (NULL == msg_data || NULL == digest_out || 0 == msg_data_size) {
     return false;
@@ -167,7 +167,7 @@ static bool evm_get_personal_data_digest(uint8_t *msg_data,
   char size_string[256] = {0};
   uint8_t size_string_size = 0;
   uint8_t *data = NULL;
-  uint16_t data_size = 0, offset = 0;
+  uint32_t data_size = 0, offset = 0;
 
   size_string_size =
       snprintf(size_string, sizeof(size_string), "%d", msg_data_size);
@@ -220,7 +220,7 @@ bool evm_get_msg_data_hash(const evm_sign_msg_context_t *ctx, uint8_t *digest) {
     case EVM_SIGN_MSG_TYPE_ETH_SIGN:
     case EVM_SIGN_MSG_TYPE_PERSONAL_SIGN: {
       result = evm_get_personal_data_digest(
-          ctx->msg_data, (uint16_t)ctx->init.total_msg_size, digest);
+          ctx->msg_data, ctx->init.total_msg_size, digest);
     } break;
 
     case EVM_SIGN_MSG_TYPE_SIGN_TYPED_DATA:
